Ticket #000 move unit tests next to code
