+++ app/vitest.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
-    include: ['test/**/*.ts'],
-    exclude: ['e2e/**']
+    include: ['**/*.test.ts'],
+    exclude: ['e2e/**', 'node_modules/**']
   }
 })
